moving methods to omnibar continued
