+++ app/src/main/java/com/duckduckgo/app/browser/omnibar/LegacyOmnibarView.kt
@@ -18,13 +18,9 @@ package com.duckduckgo.app.browser.omnibar
 
 import android.animation.Animator
 import android.content.Context
-import android.graphics.Color
-import android.graphics.drawable.ColorDrawable
-import android.text.Editable
 import android.util.AttributeSet
 import android.view.View
 import android.view.ViewGroup
-import android.widget.EditText
 import android.widget.FrameLayout
 import android.widget.ImageView
 import android.widget.ProgressBar
@@ -48,19 +44,10 @@ import com.duckduckgo.app.browser.omnibar.animations.PrivacyShieldAnimationHelpe
 import com.duckduckgo.app.browser.omnibar.model.OmnibarPosition
 import com.duckduckgo.app.browser.viewstate.BrowserViewState
 import com.duckduckgo.app.browser.viewstate.HighlightableButton
-import com.duckduckgo.app.browser.viewstate.OmnibarViewState
 import com.duckduckgo.app.global.model.PrivacyShield
-import com.duckduckgo.app.global.view.isDifferent
 import com.duckduckgo.app.statistics.pixels.Pixel
-import com.duckduckgo.app.tabs.model.TabEntity
 import com.duckduckgo.app.trackerdetection.model.Entity
-import com.duckduckgo.common.ui.DuckDuckGoActivity
 import com.duckduckgo.common.ui.view.KeyboardAwareEditText
-import com.duckduckgo.common.ui.view.KeyboardAwareEditText.ShowSuggestionsListener
-import com.duckduckgo.common.ui.view.hide
-import com.duckduckgo.common.ui.view.show
-import com.duckduckgo.common.ui.view.text.TextChangedWatcher
-import com.duckduckgo.common.utils.extractDomain
 import com.duckduckgo.di.scopes.FragmentScope
 import com.google.android.material.appbar.AppBarLayout
 import dagger.android.support.AndroidSupportInjection
@@ -176,102 +163,6 @@ class LegacyOmnibarView @JvmOverloads constructor(
         }
     }
 
-    fun getOmnibarText(): String {
-        return omnibarTextInput.text.toString()
-    }
-
-    fun isPulseAnimationPlaying(): Boolean {
-        return if (this::pulseAnimation.isInitialized) {
-            pulseAnimation.isActive
-        } else {
-            false
-        }
-    }
-
-    fun showOutline(pressed: Boolean) {
-        omniBarContainer.isPressed = pressed
-    }
-
-    fun onNewProgress(
-        newProgress: Int,
-        onAnimationEnd: (Animator?) -> Unit,
-    ) {
-        smoothProgressAnimator.onNewProgress(newProgress, onAnimationEnd)
-    }
-
-    fun addTextChangedListeners(
-        onFindInPageTextChanged: (String) -> Unit,
-        onOmnibarTextChanged: (OmnibarTextState) -> Unit,
-        onShowSuggestions: (OmnibarTextState) -> Unit,
-    ) {
-        findInPage.findInPageInput.replaceTextChangedListener(
-            object : TextChangedWatcher() {
-                override fun afterTextChanged(editable: Editable) {
-                    onFindInPageTextChanged(findInPage.findInPageInput.text.toString())
-                }
-            },
-        )
-
-        omnibarTextInput.replaceTextChangedListener(
-            object : TextChangedWatcher() {
-                override fun afterTextChanged(editable: Editable) {
-                    onOmnibarTextChanged(
-                        OmnibarTextState(
-                            omnibarTextInput.text.toString(),
-                            omnibarTextInput.hasFocus(),
-                        ),
-                    )
-                }
-            },
-        )
-
-        omnibarTextInput.showSuggestionsListener = object : ShowSuggestionsListener {
-            override fun showSuggestions() {
-                onShowSuggestions(
-                    OmnibarTextState(
-                        omnibarTextInput.text.toString(),
-                        omnibarTextInput.hasFocus(),
-                    ),
-                )
-            }
-        }
-    }
-
-    fun setScrollingEnabled(enabled: Boolean) {
-        if (isAttachedToWindow) {
-            if (enabled) {
-                omnibarScrolling.enableOmnibarScrolling(toolbarContainer)
-            } else {
-                omnibarScrolling.disableOmnibarScrolling(toolbarContainer)
-            }
-        }
-    }
-
-    fun createCookiesAnimation(isCosmetic: Boolean) {
-        animatorHelper.createCookiesAnimation(
-            context,
-            omnibarViews(),
-            cookieDummyView,
-            cookieAnimation,
-            sceneRoot,
-            isCosmetic,
-        )
-    }
-
-    fun cancelTrackersAnimation() {
-        animatorHelper.cancelAnimations(omnibarViews())
-    }
-
-    fun startTrackersAnimation(events: List<Entity>?) {
-        animatorHelper.startTrackersAnimation(
-            context = context,
-            shieldAnimationView = shieldIcon,
-            trackersAnimationView = trackersAnimation,
-            omnibarViews = omnibarViews(),
-            entities = events,
-        )
-    }
-
     fun setPrivacyShield(
         isCustomTab: Boolean,
         privacyShield: PrivacyShield,
@@ -282,128 +173,6 @@ class LegacyOmnibarView @JvmOverloads constructor(
             shieldIcon
         }
         privacyShieldView.setAnimationView(animationViewHolder, privacyShield)
-    }
-
-    fun configureCustomTab(
-        customTabToolbarColor: Int,
-        customTabDomainText: String?,
-        onTabClosePressed: () -> Unit,
-        onPrivacyShieldPressed: () -> Unit,
-    ) {
-        omniBarContainer.hide()
-        fireIconMenu.hide()
-        tabsMenu.hide()
-
-        toolbar.background = ColorDrawable(customTabToolbarColor)
-        toolbarContainer.background = ColorDrawable(customTabToolbarColor)
-
-        customTabToolbarContainer.customTabToolbar.show()
-
-        customTabToolbarContainer.customTabCloseIcon.setOnClickListener {
-            onTabClosePressed()
-        }
-
-        customTabToolbarContainer.customTabShieldIcon.setOnClickListener { _ ->
-            onPrivacyShieldPressed()
-        }
-
-        customTabToolbarContainer.customTabDomain.text = customTabDomainText
-        customTabToolbarContainer.customTabDomainOnly.text = customTabDomainText
-        customTabToolbarContainer.customTabDomainOnly.show()
-
-        val foregroundColor = calculateBlackOrWhite(customTabToolbarColor)
-        customTabToolbarContainer.customTabCloseIcon.setColorFilter(foregroundColor)
-        customTabToolbarContainer.customTabDomain.setTextColor(foregroundColor)
-        customTabToolbarContainer.customTabDomainOnly.setTextColor(
-            foregroundColor,
-        )
-        customTabToolbarContainer.customTabTitle.setTextColor(foregroundColor)
-        browserMenuImageView.setColorFilter(foregroundColor)
-    }
-
-    private fun calculateBlackOrWhite(color: Int): Int {
-        // Handle the case where we did not receive a color.
-        if (color == 0) {
-            return if ((context as DuckDuckGoActivity).isDarkThemeEnabled()) Color.WHITE else Color.BLACK
-        }
-
-        if (color == Color.WHITE || Color.alpha(color) < 128) {
-            return Color.BLACK
-        }
-        val greyValue =
-            (0.299 * Color.red(color) + 0.587 * Color.green(color) + 0.114 * Color.blue(color)).toInt()
-        return if (greyValue < 186) {
-            Color.WHITE
-        } else {
-            Color.BLACK
-        }
-    }
-
-    fun showWebPageTitleInCustomTab(
-        title: String,
-        url: String?,
-    ) {
-        customTabToolbarContainer.customTabTitle.text = title
-
-        val redirectedDomain = url?.extractDomain()
-        redirectedDomain?.let {
-            customTabToolbarContainer.customTabDomain.text = redirectedDomain
-        }
-
-        customTabToolbarContainer.customTabTitle.show()
-        customTabToolbarContainer.customTabDomainOnly.hide()
-        customTabToolbarContainer.customTabDomain.show()
-    }
-
-    fun renderOmnibarViewState(viewState: OmnibarViewState) {
-        if (viewState.navigationChange) {
-            setExpanded(true, true)
-        } else if (shouldUpdateOmnibarTextInput(viewState, viewState.omnibarText)) {
-            setOmnibarText(viewState.omnibarText)
-            if (viewState.forceExpand) {
-                setExpanded(true, true)
-            }
-            if (viewState.shouldMoveCaretToEnd) {
-                setOmnibarTextSelection(viewState.omnibarText.length)
-            }
-        }
-    }
-
-    private fun shouldUpdateOmnibarTextInput(
-        viewState: OmnibarViewState,
-        omnibarInput: String?,
-    ) =
-        (!viewState.isEditing || omnibarInput.isNullOrEmpty()) && omnibarTextInput.isDifferent(
-            omnibarInput,
-        )
-
-    fun setOmnibarText(text: String) {
-        omnibarTextInput.setText(text)
-    }
-
-    fun setOmnibarTextSelection(index: Int) {
-        omnibarTextInput.setSelection(index)
-    }
-
-    fun renderVoiceSearch(
-        viewState: BrowserViewState,
-        voiceSearchPressed: () -> Unit,
-    ) {
-        if (viewState.showVoiceSearch) {
-            voiceSearchButton.visibility = VISIBLE
-            voiceSearchButton.setOnClickListener {
-                voiceSearchPressed()
-            }
-        } else {
-            voiceSearchButton.visibility = GONE
-        }
-    }
-
-    fun showOmnibarTextSpacer(
-        showClearButton: Boolean,
-        showVoiceSearch: Boolean,
-    ) {
-        spacer.isVisible = showVoiceSearch && showClearButton
     }
 
     fun renderToolbarButtons(
@@ -432,7 +201,17 @@ class LegacyOmnibarView @JvmOverloads constructor(
         renderPulseAnimation(viewState)
     }
 
-    private fun renderPulseAnimation(viewState: BrowserViewState) {
+    fun setScrollingEnabled(enabled: Boolean) {
+        if (isAttachedToWindow) {
+            if (enabled) {
+                omnibarScrolling.enableOmnibarScrolling(toolbarContainer)
+            } else {
+                omnibarScrolling.disableOmnibarScrolling(toolbarContainer)
+            }
+        }
+    }
+
+    fun renderPulseAnimation(viewState: BrowserViewState) {
         val targetView = if (viewState.showMenuButton.isHighlighted()) {
             browserMenuImageView
         } else if (viewState.fireButton.isHighlighted()) {
@@ -457,25 +236,43 @@ class LegacyOmnibarView @JvmOverloads constructor(
         }
     }
 
-    fun animateTabsCount() {
-        tabsMenu.animateCount()
-    }
-
-    fun renderTabIcon(tabs: List<TabEntity>) {
-        context?.let {
-            tabsMenu.count = tabs.count()
-            tabsMenu.hasUnread = tabs.firstOrNull { !it.viewed } != null
+    fun isPulseAnimationPlaying(): Boolean {
+        return if (this::pulseAnimation.isInitialized) {
+            pulseAnimation.isActive
+        } else {
+            false
         }
     }
 
-    fun incrementTabs(onTabsIncremented: () -> Unit) {
-        tabsMenu.increment {
-            onTabsIncremented()
-        }
+    fun createCookiesAnimation(isCosmetic: Boolean) {
+        animatorHelper.createCookiesAnimation(
+            context,
+            omnibarViews(),
+            cookieDummyView,
+            cookieAnimation,
+            sceneRoot,
+            isCosmetic,
+        )
     }
-}
 
-private fun EditText.replaceTextChangedListener(textWatcher: TextChangedWatcher) {
-    removeTextChangedListener(textWatcher)
-    addTextChangedListener(textWatcher)
+    fun cancelTrackersAnimation() {
+        animatorHelper.cancelAnimations(omnibarViews())
+    }
+
+    fun startTrackersAnimation(events: List<Entity>?) {
+        animatorHelper.startTrackersAnimation(
+            context = context,
+            shieldAnimationView = shieldIcon,
+            trackersAnimationView = trackersAnimation,
+            omnibarViews = omnibarViews(),
+            entities = events,
+        )
+    }
+
+    fun onNewProgress(
+        newProgress: Int,
+        onAnimationEnd: (Animator?) -> Unit,
+    ) {
+        smoothProgressAnimator.onNewProgress(newProgress, onAnimationEnd)
+    }
 }
